requirements.txt
flask
flask_cors
python-dotenv
werkzeug.utils
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
pdfplumber
nest_asyncio